requirements.txt
gdown
matplotlib
numpy
opencv_python
pandas
pillow
torch
torchvision
tqdm
lightning
wandb
